requirements.txt
streamlit
requests
python-dotenv
flask
groq